requirements.txt
transformers
accelerate
django
djangorestframework
